requirements.txt
streamlit
torch
st-annotated-text
extra-streamlit-components
streamlit-extras
